+++ EssentialFeed/EssentialFeediOS/Feed UI/Composers/FeedUIComposer.swift
@@ -37,26 +37,6 @@ private extension FeedViewController {
 	}
 }
 
-private final class WeakRefVirtualProxy<T: AnyObject> {
-	private weak var object: T?
-	
-	init(_ object: T) {
-		self.object = object
-	}
-}
-
-extension WeakRefVirtualProxy: FeedLoadingView where T: FeedLoadingView {
-	func display(_ viewModel: FeedLoadingViewModel) {
-		object?.display(viewModel)
-	}
-}
-
-extension WeakRefVirtualProxy: FeedImageView where T: FeedImageView, T.Image == UIImage {
-	func display(_ model: FeedImageViewModel<UIImage>) {
-		object?.display(model)
-	}
-}
-
 private final class FeedViewAdapter: FeedView {
 	private weak var controller: FeedViewController?
 	private let imageLoader: FeedImageDataLoader
